# docs/swagger.yaml
definitions:
  models.GFSAction:
    enum:
    - GFSForceLoad
    - GFSDistributeLoad
    - GFSLoadMetadata
    - GFSList
    type: string
    x-enum-comments:
      GFSDistributeLoad: 缓存数据，他的依据是 Master 上的 metadata
      GFSForceLoad: 该步骤执行的是先去 LoadMetadata，然后再去 DistributeLoad，这样彻底更新
      GFSLoadMetadata: 只更新元数据信息，可以更新掉cos上变更的内容
    x-enum-varnames:
    - GFSForceLoad
    - GFSDistributeLoad
    - GFSLoadMetadata
    - GFSList
  models.GooseFSRequest:
    properties:
      action:
        allOf:
        - $ref: '#/definitions/models.GFSAction'
        description: '必填 0: GooseFSDistributeLoad 1: GooseFSLoadMetadata 2: GooseFSList
          3: GooseFSForceLoad'
      path:
        description: 当 action 为 GooseFSDistributeLoad/GooseFSLoadMetadata/GooseFSList
          时必填
        items:
          type: string
        type: array
      task_name:
        description: 选填，支持提交多个任务到同一个任务标签上
        type: string
      timeout:
        description: 当 action 为 GooseFSList 由于没有挂起任务，所以需要指定超时时间 默认 30 秒
        type: integer
    required:
    - action
    - path
    type: object
  models.QueryTaskRequest:
    properties:
      task_id:
        type: string
      task_name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  termsOfService: http://swagger.io/terms/
  title: GooseFS-CLI2API
  version: v1
paths:
  /api/v1/gfs:
    post:
      consumes:
      - application/json
      description: 执行内置的 goosefs 命令，包括 distribute_load/load_metadata，返回 task_id，可以通过
        task_id 获取执行状态或者输出;注意GooseFSList是等待执行的不是挂起的任务，且只支持 1 个 path查询。
      parameters:
      - description: DistrubuteLoad
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.GooseFSRequest'
      produces:
      - application/json
      responses:
        "200":
          description: task_id
          schema:
            type: string
      summary: execute for goosefs cli
      tags:
      - GooseFS
  /api/v1/gfs/report:
    get:
      consumes:
      - application/json
      description: GooseFSReport 获取 goosefs 集群状态
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: GooseFSReport
      tags:
      - GooseFS
  /api/v1/output:
    get:
      consumes:
      - application/json
      description: GetTaskOutput
      parameters:
      - description: task_id
        in: query
        name: task_id
        required: true
        type: string
      - description: task_name
        in: query
        name: task_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: GetTaskOutput
  /api/v1/status:
    get:
      consumes:
      - application/json
      description: GetTaskStatus
      parameters:
      - description: task_id
        in: query
        name: task_id
        required: true
        type: string
      - description: task_name
        in: query
        name: task_name
        type: string
      - description: action
        in: query
        name: action
        type: string
      - description: status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QueryTaskRequest'
      summary: GetTaskStatus
swagger: "2.0"